+++ test/metar-parser.js
@@ -12,11 +12,11 @@ describe('metarParser', function() {
       metarCode: "KEYW 041053Z AUTO 13005KT 10SM CLR 24/22 A3000 RMK AO2 SLP159 T02440222",
       expectedValues: [
         ['icao', 'KEYW'],
-        ['wind', {degrees: 130, speed_kt: 5, gust_kt: 5}],
-        ['visibility_m', 10 * 1609],
+        ['wind', {degrees: 130, speed_kts: 5, gust_kts: 5}],
+        ['visibility', { meters: 10 * 1609 }],
         ['clouds', []],
-        ['temperature_c', 24],
-        ['barometer_hpa', 3000 / 2.9529988]
+        ['temperature', { celsius: 24 }],
+        ['barometer', {kpa: 3000 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -25,11 +25,11 @@ describe('metarParser', function() {
       metarCode: "KACV 041053Z AUTO 07003KT 10SM CLR 04/04 A3001 RMK AO2 SLP169 T00440039",
       expectedValues: [
         ['icao', 'KACV'],
-        ['wind', {degrees: 70, speed_kt: 3, gust_kt: 3}],
-        ['visibility_m', 10 * 1609],
+        ['wind', {degrees: 70, speed_kts: 3, gust_kts: 3}],
+        ['visibility', { meters: 10 * 1609 }],
         ['clouds', []],
-        ['temperature_c', 4],
-        ['barometer_hpa', 3001 / 2.9529988]
+        ['temperature', { celsius: 4 }],
+        ['barometer', {kpa: 3001 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -38,11 +38,11 @@ describe('metarParser', function() {
       metarCode: "KPIE 260853Z AUTO 02013G17KT 10SM CLR 17/07 A2998 RMK AO2 SLP153 T01720072 57000",
       expectedValues: [
         ['icao', 'KPIE'],
-        ['wind', {degrees: 20, speed_kt: 13, gust_kt: 17}],
-        ['visibility_m', 10 * 1609],
+        ['wind', {degrees: 20, speed_kts: 13, gust_kts: 17}],
+        ['visibility', { meters: 10 * 1609 }],
         ['clouds', []],
-        ['temperature_c', 17],
-        ['barometer_hpa', 2998 / 2.9529988]
+        ['temperature', { celsius: 17 }],
+        ['barometer', {kpa: 2998 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -51,11 +51,11 @@ describe('metarParser', function() {
       metarCode: "KSPG 260853Z AUTO 05012KT 10SM CLR 18/09 A2997 RMK AO2 SLP148 T01830094 53001",
       expectedValues: [
         ['icao', 'KSPG'],
-        ['wind', {degrees: 50, speed_kt: 12, gust_kt: 12}],
-        ['visibility_m', 10 * 1609],
+        ['wind', {degrees: 50, speed_kts: 12, gust_kts: 12}],
+        ['visibility', { meters: 10 * 1609 }],
         ['clouds', []],
-        ['temperature_c', 18],
-        ['barometer_hpa', 2997 / 2.9529988]
+        ['temperature', { celsius: 18 }],
+        ['barometer', {kpa: 2997 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -64,11 +64,11 @@ describe('metarParser', function() {
       metarCode: "EDDS 081620Z 29010KT 9999 FEW040TCU 09/M03 Q1012 NOSIG",
       expectedValues: [
         ['icao', 'EDDS'],
-        ['wind', {degrees: 290, speed_kt: 10, gust_kt: 10}],
-        ['visibility_m', 9999],
-        ['clouds', [{baseAgl_ft: 4000, minDensity: 1, maxDensity: 2}]],
-        ['temperature_c', 9],
-        ['barometer_hpa', 1012]
+        ['wind', {degrees: 290, speed_kts: 10, gust_kts: 10}],
+        ['visibility', { meters: 9999 }],
+        ['clouds', [{base_feet_agl: 4000, minDensity: 1, maxDensity: 2}]],
+        ['temperature', { celsius: 9 }],
+        ['barometer', {kpa: 101.2 }]
       ],
       output: false
     },
@@ -77,11 +77,11 @@ describe('metarParser', function() {
       metarCode: "METAR LBBG 041600Z 12012MPS 090V150 1400 R04/P1500N R22/P1500U +SN BKN022 OVC050 M04/M07 Q1020 NOSIG 8849//91=",
       expectedValues: [
         ['icao', 'LBBG'],
-        ['wind', {degrees: 120, speed_kt: 12 / 1.9438445, gust_kt: 12 / 1.9438445}],
-        ['visibility_m', 1400],
-        ['clouds', [{baseAgl_ft: 2200, minDensity: 5, maxDensity: 7}, {baseAgl_ft: 5000, minDensity: 8, maxDensity: 8}]],
-        ['temperature_c', -4],
-        ['barometer_hpa', 1020]
+        ['wind', {degrees: 120, speed_kts: 12 / 1.9438445, gust_kts: 12 / 1.9438445}],
+        ['visibility', { meters: 1400 }],
+        ['clouds', [{base_feet_agl: 2200, minDensity: 5, maxDensity: 7}, {base_feet_agl: 5000, minDensity: 8, maxDensity: 8}]],
+        ['temperature', { celsius: -4 }],
+        ['barometer', {kpa: 102.0 }]
       ],
       output: false
     },
@@ -90,11 +90,11 @@ describe('metarParser', function() {
       metarCode: "METAR KTTN 051853Z 04011KT 1/2SM VCTS SN FZFG BKN003 OVC010 M02/M02 A3006 RMK AO2 TSB40 SLP176 P0002 T10171017=",
       expectedValues: [
         ['icao', 'KTTN'],
-        ['wind', {degrees: 40, speed_kt: 11, gust_kt: 11}],
-        ['visibility_m', 0.5 * 1609],
-        ['clouds', [{baseAgl_ft: 300, minDensity: 5, maxDensity: 7}, {baseAgl_ft: 1000, minDensity: 8, maxDensity: 8}]],
-        ['temperature_c', -2],
-        ['barometer_hpa', 3006 / 2.9529988]
+        ['wind', {degrees: 40, speed_kts: 11, gust_kts: 11}],
+        ['visibility', { meters: 0.5 * 1609 }],
+        ['clouds', [{base_feet_agl: 300, minDensity: 5, maxDensity: 7}, {base_feet_agl: 1000, minDensity: 8, maxDensity: 8}]],
+        ['temperature', { celsius: -2 }],
+        ['barometer', {kpa: 3006 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -103,11 +103,11 @@ describe('metarParser', function() {
       metarCode: "KEYW 041053Z AUTO 13005KT 10SM CLR 24/22 A3000 RMK AO2 SLP159 T02440222",
       expectedValues: [
         ['icao', 'KEYW'],
-        ['wind', {degrees: 130, speed_kt: 5, gust_kt: 5}],
-        ['visibility_m', 10 * 1609],
+        ['wind', {degrees: 130, speed_kts: 5, gust_kts: 5}],
+        ['visibility', { meters: 10 * 1609 }],
         ['clouds', []],
-        ['temperature_c', 24],
-        ['barometer_hpa', 3000 / 2.9529988]
+        ['temperature', { celsius: 24 }],
+        ['barometer', {kpa: 3000 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -116,11 +116,11 @@ describe('metarParser', function() {
       metarCode: "EDDH 041050Z 29013KT 6000 SCT006 BKN009 04/03 Q1028 TEMPO BKN012",
       expectedValues: [
         ['icao', 'EDDH'],
-        ['wind', {degrees: 290, speed_kt: 13, gust_kt: 13}],
-        ['visibility_m', 6000],
-        ['clouds', [{baseAgl_ft: 600, minDensity: 3, maxDensity: 4}, {baseAgl_ft: 900, minDensity: 5, maxDensity: 7}]],
-        ['temperature_c', 4],
-        ['barometer_hpa', 1028]
+        ['wind', {degrees: 290, speed_kts: 13, gust_kts: 13}],
+        ['visibility', { meters: 6000 }],
+        ['clouds', [{base_feet_agl: 600, minDensity: 3, maxDensity: 4}, {base_feet_agl: 900, minDensity: 5, maxDensity: 7}]],
+        ['temperature', { celsius: 4 }],
+        ['barometer', {kpa: 102.8 }]
       ],
       output: false
     },
@@ -129,11 +129,11 @@ describe('metarParser', function() {
       metarCode: "ETEB 041056Z AUTO 26010KT 9999 SCT090 00/M01 A3052 RMK AO2 SLP378 T10031013",
       expectedValues: [
         ['icao', 'ETEB'],
-        ['wind', {degrees: 260, speed_kt: 10, gust_kt: 10}],
-        ['visibility_m', 9999],
-        ['clouds', [{baseAgl_ft: 9000, minDensity: 3, maxDensity: 4}]],
-        ['temperature_c', 0],
-        ['barometer_hpa', 3052 / 2.9529988]
+        ['wind', {degrees: 260, speed_kts: 10, gust_kts: 10}],
+        ['visibility', { meters: 9999 }],
+        ['clouds', [{base_feet_agl: 9000, minDensity: 3, maxDensity: 4}]],
+        ['temperature', { celsius: 0 }],
+        ['barometer', {kpa: 3052 / 10 / 2.9529988 }]
       ],
       output: false
     },
@@ -142,11 +142,19 @@ describe('metarParser', function() {
       metarCode: "KEYW 050653Z AUTO 19006KT FEW024 BKN039 26/23 A3000 RMK AO2 LTG DSNT W SLP159 T02610228",
       expectedValues: [
         ['icao', 'KEYW'],
-        ['wind', {degrees: 190, speed_kt: 6, gust_kt: 6}],
-        ['visibility_m', 9999],
-        ['clouds', [{baseAgl_ft: 2400, minDensity: 1, maxDensity: 2}, {baseAgl_ft: 3900, minDensity: 5, maxDensity: 7}]],
-        ['temperature_c', 26],
-        ['barometer_hpa', 3000 / 2.9529988]
+        ['wind', {degrees: 190, speed_kts: 6, gust_kts: 6}],
+        ['visibility', { meters: 9999 }],
+        ['clouds', [{base_feet_agl: 2400, minDensity: 1, maxDensity: 2}, {base_feet_agl: 3900, minDensity: 5, maxDensity: 7}]],
+        ['temperature', { celsius: 26 }],
+        ['barometer', {kpa: 3000 / 10 / 2.9529988 }]
+      ],
+      output: false
+    },
+    {
+      source: 'https://api.checkwx.com/#2019-01-07',
+      metarCode: 'KSFO 070121Z 19023KT 1 1/2SM R28R/6000VP6000FT -RA BKN004 BKN013 OVC035 15/12 A2970 RMK AO2 T01500122 PNO $',
+      expectedValues: [
+        ['visibility', { meters: 1.5 * 1609 }]
       ],
       output: false
     }