+++ test/metar-to-aerofly.js
@@ -10,13 +10,17 @@ describe('metarToAerofly', function() {
     const aeroflyObject = metarToAerofly({
       icao: 'KPIE',
       observed: new Date('2019-01-29T12:30:18.825Z'),
-      wind: { degrees: 270, speed_kt: 21, gust_kt: 30 },
-      visibility_m: 20000,
+      wind: { degrees: 270, speed_kts: 21, gust_kts: 30 },
+      visibility: { meters: 20000 },
       clouds: [
-        { baseAgl_ft: 2900, minDensity: 5, maxDensity: 7 }
+        { base_feet_agl: 2900, minDensity: 5, maxDensity: 7 }
       ],
-      temperature_c: 18,
-      barometer_hpa: 1010
+      temperature: {
+        celsius: 18
+      },
+      barometer: {
+        kpa: 101.0
+      }
     });
 
     // console.log(aeroflyObject);
@@ -24,22 +28,26 @@ describe('metarToAerofly', function() {
     assert.strictEqual(aeroflyObject.time_day,      29);
     assert.strictEqual(aeroflyObject.time_month,    1);
     assert.strictEqual(aeroflyObject.time_hours,    12.5);
-    assert.strictEqual(aeroflyObject.wind_direction_in_degree,  270);
-    assert.strictEqual(aeroflyObject.visibility,    1);
-    assert.strictEqual(aeroflyObject.clouds.length, 1);
+    assert.strictEqual(aeroflyObject.wind_direction_in_degree,  270, 'Wind direction');
+    assert.strictEqual(aeroflyObject.visibility,    1, 'Visibility');
+    assert.strictEqual(aeroflyObject.clouds.length, 1, 'Clouds');
   });
 
   it('should convert METAR object to Aerofly object #2', function() {
     const aeroflyObject = metarToAerofly({
       icao: 'KPIE',
       observed: new Date('2019-01-29T12:30:18.825Z'),
-      wind: { degrees: 270, speed_kt: 20, gust_kt: 20 },
-      visibility_m: 20000,
+      wind: { degrees: 270, speed_kts: 20, gust_kts: 20 },
+      visibility: { meters: 20000 },
       clouds: [
-        { baseAgl_ft: 2900, minDensity: 5, maxDensity: 7 }
+        { base_feet_agl: 2900, minDensity: 5, maxDensity: 7 }
       ],
-      temperature_c: 18,
-      barometer_hpa: 1010
+      temperature: {
+        celsius: 18
+      },
+      barometer: {
+        kpa: 101.0
+      }
     });
 
     // console.log(aeroflyObject);
@@ -56,13 +64,17 @@ describe('metarToAerofly', function() {
     const aeroflyObject = metarToAerofly({
       icao: 'KPIE',
       observed: new Date('2019-01-29T18:30:18.825Z'),
-      wind: { degrees: 270, speed_kt: 5, gust_kt: 5 },
-      visibility_m: 20000,
+      wind: { degrees: 270, speed_kts: 5, gust_kts: 5 },
+      visibility: { meters: 20000 },
       clouds: [
-        { baseAgl_ft: 2900, minDensity: 5, maxDensity: 7 }
+        { base_feet_agl: 2900, minDensity: 5, maxDensity: 7 }
       ],
-      temperature_c: 18,
-      barometer_hpa: 1010
+      temperature: {
+        celsius: 18
+      },
+      barometer: {
+        kpa: 101.0
+      }
     }, {
       hourOffset: 12
     });